# bootcamp/05-devops/testing/demo/openAPI.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterDto" }
            }
          }
        },
        "responses": {
          "409": { "description": "That email address is already registered." }
        },
        "summary": "Create an account and receive a token",
        "tags": ["auth"]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginDto" }
            }
          }
        },
        "responses": {
          "401": { "description": "Email or password did not match." }
        },
        "summary": "Exchange credentials for a token",
        "tags": ["auth"]
      }
    },
...
  },
  "info": {
    "title": "DarkBay API",
    "description": "An auction API. Browsing is public; listing an item and placing a bid require a bearer token from `POST /auth/login`.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste the `accessToken` returned by /auth/login."
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email address used to log in. Must not already be registered.",
            "format": "email",
            "example": "seller@darkbay.dev"
          },
          "password": {
            "type": "string",
            "description": "Plaintext password. Stored only as a bcrypt hash.",
            "minLength": 8,
            "maxLength": 72,
            "example": "correct-horse-battery"
          }
        },
        "required": ["email", "password"]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "seller@darkbay.dev"
          },
          "password": { "type": "string", "example": "correct-horse-battery" }
        },
        "required": ["email", "password"]
      },
...
    }
  }
}